refactor: MemberService 분리 #144
