+++ src/main/java/SMU/BAMBOO/Hompage/domain/member/service/MemberService.java
@@ -9,15 +9,31 @@ import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 
 public interface MemberService {
 
-    MemberResponseDTO.MemberInfo signUp(MemberRequestDTO.SignUp request, BCryptPasswordEncoder encoder);
-    MemberResponseDTO.Login login(MemberRequestDTO.Login request, HttpServletResponse response);
-    Member getMember(String studentId);
-    Page<MemberResponseDTO.MemberInfo> getMembers(int page, int size);
-    String logout(String accessToken);
-    MemberResponseDTO.MyPage updateProfile(Long memberId, MemberRequestDTO.UpdateProfile request);
-    MemberResponseDTO.MyPage deleteProfileImage(Long memberId);
-    void updatePw(Long memberId, MemberRequestDTO.UpdatePw request);
-    void resetPw(MemberRequestDTO.ResetPw request);
-    MemberResponseDTO.MemberInfo updateRole(Long currentMemberId, MemberRequestDTO.UpdateRole request);
-    void deactivateMember(Long memberId);
+    interface AuthenticationService {
+        MemberResponseDTO.MemberInfo signUp(MemberRequestDTO.SignUp request, BCryptPasswordEncoder encoder);
+
+        MemberResponseDTO.Login login(MemberRequestDTO.Login request, HttpServletResponse response);
+
+        String logout(String accessToken);
+    }
+
+    interface MemberInfoService {
+        Member getMember(String studentId);
+
+        Page<MemberResponseDTO.MemberInfo> getMembers(int page, int size);
+
+        MemberResponseDTO.MyPage updateProfile(Long memberId, MemberRequestDTO.UpdateProfile request);
+
+        MemberResponseDTO.MyPage deleteProfileImage(Long memberId);
+
+        void updatePw(Long memberId, MemberRequestDTO.UpdatePw request);
+
+        void resetPw(MemberRequestDTO.ResetPw request);
+    }
+
+    interface AdminService {
+        MemberResponseDTO.MemberInfo updateRole(Long currentMemberId, MemberRequestDTO.UpdateRole request);
+
+        void deactivateMember(Long memberId);
+    }
 }
